==> projekt/3_Random_Forest.Rmd <==
---
title: "Random Forest model development for Scorecards project"
author: "Zuzanna Kostecka, Karolina Solarska"
date: "2023-06-30"
output: html_document
---

```{r}
# Karolina:
# setwd("C:/Users/karla/OneDrive/Pulpit/Credit Risk/Projekt")
# app_train5<-read_csv("app_test4.csv")
# app_test4<-read_csv("app_train5.csv")

# Zuzia:
setwd("/Users/zuzanna/Desktop/Studia/CreditRisk/projekt")
load("dane/app_train4.RData")
load("dane/app_test4.RData")

```

```{r}
library(randomForest)
library(caret)
```

<br>
<br>

# 1. Preparing datasets


```{r}

app_train5 <- app_train4[c(1:47511),]
app_train5$TARGET = as.factor(app_train5$TARGET)

trainIndex <- createDataPartition(app_train5$TARGET, p = .70,
                                  list = FALSE,
                                  times = 1)

# Split the data based on the partition
Trainsample <- app_train5[trainIndex, ]
Validsample <- app_train5[-trainIndex, ]
```

<br>

# 2. Model training

```{r}
model_rf1 <- randomForest(TARGET ~ NAME_CONTRACT_TYPE + CODE_GENDER + FLAG_OWN_CAR + 
                      FLAG_OWN_REALTY + CNT_CHILDREN + AMT_INCOME_TOTAL + 
                      AMT_CREDIT + AMT_ANNUITY + AMT_GOODS_PRICE + NAME_TYPE_SUITE + 
                      NAME_EDUCATION_TYPE + NAME_FAMILY_STATUS + DAYS_EMPLOYED + 
                      REGION_RATING_CLIENT + DOCUMENT + EXT_SOURCE_1 + 
                      EXT_SOURCE_2 + EXT_SOURCE_3 + AMT_REQ_CREDIT_BUREAU_QRT, 
                    data = Trainsample, importance = TRUE)

model_rf1
```

<br>

# 3. Model validation

```{r}
predictedvalid_2 <- predict(model_rf1, Validsample, type = "prob")
Validsample_rf <- data.frame('SK_ID_CURR' = Validsample$SK_ID_CURR, 
                             'Predictedvalues' = predictedvalid_2)
Validsample_rf <- as.data.frame(Validsample_rf)
predictedvalid_2 <- as.data.frame(predictedvalid_2)

Validsample_rf[,4] <- ifelse(Validsample_rf$Predictedvalues.1 > 0.4, 1, 0)

Validsample_rf[,4] <- as.factor(Validsample_rf[,4])
Validsample$TARGET <- as.factor(Validsample$TARGET)

summary(Validsample_rf[,4])
summary(Validsample$TARGET)
```
<br>

```{r}
confusionMatrix(Validsample_rf$V4, Validsample$TARGET)
```

<br>

# 4. Test sample prediction

```{r}
predictedvalues_test <- predict(model_rf1, app_test4, type = "prob")
Testsamplerf <- data.frame('SK_ID_CURR' = app_test4$SK_ID_CURR, 
                            'Predictedvalues' = predictedvalues_test)

Testsamplerf <- as.data.frame(Testsamplerf)
Testsamplerf[,4] <- ifelse(Testsamplerf$Predictedvalues.1 > 0.4, 1, 0)
```

```{r}
colnames(Testsamplerf)[colnames(Testsamplerf) == "V4"] <- "Predictedvalues"
```

```{r}
save(Testsamplerf, file="RF_pred_test.RData")
```

<br>

# 5. Assigning the scoring class

We have created a simplified scoring classification (proportional). Nevertheless,
we are aware this is not the commonly business approach. For educational and comparison
purposes, we decided to apply this simplify method.
```{r}
credit_ratings <- c("AA+", "AA", "AA-", "A+", "A", "A-", "BBB+", "BBB", "BBB-", 
                    "BB+", "BB", "BB-", "B+", "B", "B-", "CCC+", "CCC")

Testsamplerf$Credit_Rating <- cut(Testsamplerf$Predictedvalues.1, 
                                  breaks = c(0, 0.510, 0.530, 0.560, 1),
                                  labels = c(NA, "CCC-", "D+", "D"), 
                                  right = FALSE)
summary(Testsamplerf$Credit_Rating)

Testsamplerf$Credit_Rating <- cut(Testsamplerf$Predictedvalues.1, 
                                 breaks = c(-Inf, seq(0, 0.509, length.out = 17)), 
                                 labels = credit_ratings, right = FALSE)

Testsamplerf$Credit_Rating[Testsamplerf$Predictedvalues.0==1] <- "AA+"

# Identify NA values in Credit_Rating column
na_indices <- is.na(Testsamplerf$Credit_Rating)

# Generate random ratings from the available credit ratings
random_ratings <- sample(credit_ratings, sum(na_indices), replace = TRUE)

# Assign the random ratings to the NA values
Testsamplerf$Credit_Rating[na_indices] <- random_ratings


# Print the summary of the resulting dataframe
summary(Testsamplerf$Credit_Rating)
Testsamplerf[is.na(Testsamplerf$Credit_Rating),]

```

<br>
We observe high deficiency of defaulting clients (imbalanced dataset), which is not
the best case for Random Forest technique.
```{r}
targetTE <- table(Testsamplerf$Credit_Rating)
barplot(targetTE, main = "Ratings made form Random Forest",
        ylab = "Count",
        names.arg = credit_ratings,
        cex.names = 0.5)
```
